requirements.txt
attrs==22.1.0
aws-cdk-lib==2.46.0
boto3==1.24.95
botocore==1.27.95
cattrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constructs==10.1.132
coverage==6.5.0
cryptography==38.0.3
exceptiongroup==1.0.0rc9
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.69.0
MarkupSafe==2.1.1
moto==4.0.8
packaging==21.3
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
responses==0.22.0
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typeguard==2.13.3
types-toml==0.10.8
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
xmltodict==0.13.0
pydantic==1.10.4
pylint_pydantic==0.1.4